I added the monsters to the with their scripts
I fixed the materials for the crypt and the monsters. The monsters are also working with their scripts. I just need to integrate the dungeon generation with the dialogue and make sure it is all working well. I also need to make it so the character switches between the two control schemes.